+++ src/interceptor/node/client-request.ts
@@ -212,7 +212,10 @@ function ClientRequest(
           const { body, json, response } = await simpleRequest({
             url: remoteInfo.url,
             method: remoteInfo.method || this.options.method || 'GET',
-            headers: requestInfo.headers as Record<string, string>,
+            headers: {
+              ...requestInfo.headers as Record<string, string>,
+              ...mockItem.requestHeaders,
+            },
             body: this.requestBody
           });
           remoteResponse = {